Add MQCIH header
