+++ native/src/main/java/io/ballerina/lib/ibm.ibmmq/CommonUtils.java
@@ -22,20 +22,8 @@ import com.ibm.mq.MQException;
 import com.ibm.mq.MQGetMessageOptions;
 import com.ibm.mq.MQMessage;
 import com.ibm.mq.MQPropertyDescriptor;
-import com.ibm.mq.headers.MQCIH;
-import com.ibm.mq.headers.MQDH;
-import com.ibm.mq.headers.MQDLH;
-import com.ibm.mq.headers.MQDataException;
 import com.ibm.mq.headers.MQHeaderList;
-import com.ibm.mq.headers.MQIIH;
-import com.ibm.mq.headers.MQRFH;
-import com.ibm.mq.headers.MQRFH2;
-import com.ibm.mq.headers.MQRMH;
-import com.ibm.mq.headers.MQSAPH;
-import com.ibm.mq.headers.MQTM;
-import com.ibm.mq.headers.MQWIH;
-import com.ibm.mq.headers.MQXQH;
-import com.ibm.mq.headers.pcf.MQEPH;
+import io.ballerina.lib.ibm.ibmmq.headers.MQRFH2Header;
 import io.ballerina.runtime.api.PredefinedTypes;
 import io.ballerina.runtime.api.Runtime;
 import io.ballerina.runtime.api.creators.ErrorCreator;
@@ -44,25 +32,22 @@ import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BError;
-import io.ballerina.runtime.api.values.BIterator;
 import io.ballerina.runtime.api.values.BMap;
-import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.runtime.api.values.BTable;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Enumeration;
-import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.concurrent.CountDownLatch;
 
 import static io.ballerina.lib.ibm.ibmmq.Constants.IBMMQ_ERROR;
-import static io.ballerina.lib.ibm.ibmmq.Constants.MQRFH2FIELD_RECORD_NAME;
 import static io.ballerina.lib.ibm.ibmmq.ModuleUtils.getModule;
+import static io.ballerina.lib.ibm.ibmmq.headers.MQCHIHHeader.createMQCIHHeaderFromBHeader;
+import static io.ballerina.lib.ibm.ibmmq.headers.MQRFH2Header.createMQRFH2HeaderFromBHeader;
+import static io.ballerina.lib.ibm.ibmmq.headers.MQRFHHeader.createMQRFHHeaderFromBHeader;
 
 /**
  * {@code CommonUtils} contains the common utility functions for the Ballerina IBM MQ connector.
@@ -88,22 +73,6 @@ public class CommonUtils {
     private static final BString PROPERTY_DESCRIPTOR = StringUtils.fromString("descriptor");
     private static final BString WAIT_INTERVAL = StringUtils.fromString("waitInterval");
     private static final BString OPTIONS = StringUtils.fromString("options");
-    private static final BString FIELD_VALUES_FIELD = StringUtils.fromString("fieldValues");
-    private static final BString FOLDER_STRINGS_FIELD = StringUtils.fromString("folderStrings");
-    private static final BString FLAGS_FIELD = StringUtils.fromString("flags");
-    private static final BString VERSION_FIELD = StringUtils.fromString("version");
-    private static final BString NAME_VALUE_CCSID_FIELD = StringUtils.fromString("nameValueCCSID");
-    private static final BString NAME_VALUE_DATA_FIELD = StringUtils.fromString("nameValueData");
-    private static final BString STRUC_ID_FIELD = StringUtils.fromString("strucId");
-    private static final BString STRUC_LENGTH_FIELD = StringUtils.fromString("strucLength");
-    private static final BString FOLDER_FIELD = StringUtils.fromString("folder");
-    private static final BString FIELD_FIELD = StringUtils.fromString("field");
-    private static final BString VALUE_FIELD = StringUtils.fromString("value");
-    private static final BString NAME_VALUE_PAIRS_FIELD = StringUtils.fromString("nameValuePairs");
-    private static final String MQRFH2_RECORD_NAME = "MQRFH2";
-    private static final String MQRFH_RECORD_NAME = "MQRFH";
-    private static final String NATIVE_UTILS_OBJECT_NAME = "NativeUtils";
-    private static final String ADD_FIELDS_TO_TABLE_FUNCTION_NAME = "addMQRFH2FieldsToTable";
 
     private static final MQPropertyDescriptor defaultPropertyDescriptor = new MQPropertyDescriptor();
 
@@ -230,6 +199,7 @@ public class CommonUtils {
             switch (headerType) {
                 case MQRFH2 -> headerList.add(createMQRFH2HeaderFromBHeader(bHeader));
                 case MQRFH -> headerList.add(createMQRFHHeaderFromBHeader(bHeader));
+                case MQCIH -> headerList.add(createMQCIHHeaderFromBHeader(bHeader));
                 default -> throw createError(IBMMQ_ERROR,
                         String.format("Error occurred while populating headers: Unsupported header type"), null);
             }
@@ -239,72 +209,6 @@ public class CommonUtils {
         } catch (IOException e) {
             throw createError(IBMMQ_ERROR,
                     String.format("Error occurred while putting a message to the topic: %s", e.getMessage()), e);
-        }
-    }
-
-    private static MQRFH createMQRFHHeaderFromBHeader(BMap<BString, Object> bHeader) {
-        MQRFH header = new MQRFH();
-        header.setFlags(bHeader.getIntValue(FLAGS_FIELD).intValue());
-        BMap<BString, Object> nameValuePairsMap = (BMap<BString, Object>) bHeader.getMapValue(NAME_VALUE_PAIRS_FIELD);
-        for (BString key : nameValuePairsMap.getKeys()) {
-            try {
-                header.addNameValuePair(key.getValue(), nameValuePairsMap.getStringValue(key).getValue());
-            } catch (IOException e) {
-                throw createError(IBMMQ_ERROR, String
-                        .format("Error occurred while adding key pair values to MQRFH header: %s", e.getMessage()), e);
-            }
-        }
-        return header;
-    }
-
-    private static MQRFH2 createMQRFH2HeaderFromBHeader(BMap<BString, Object> bHeader) {
-        MQRFH2 header = new MQRFH2();
-        header.setFlags(bHeader.getIntValue(FLAGS_FIELD).intValue());
-        BArray folderStringsArray = bHeader.getArrayValue(FOLDER_STRINGS_FIELD);
-        try {
-            header.setFolderStrings(folderStringsArray.getStringArray());
-        } catch (IOException e) {
-            throw createError(IBMMQ_ERROR, String
-                    .format("Error occurred while setting folder string to MQRFH2 header: %s", e.getMessage()), e);
-        }
-        header.setNameValueCCSID(bHeader.getIntValue(NAME_VALUE_CCSID_FIELD).intValue());
-        header.setNameValueData(bHeader.getArrayValue(NAME_VALUE_DATA_FIELD).getBytes());
-        BTable fieldTable = (BTable) bHeader.get(FIELD_VALUES_FIELD);
-        BIterator fieldTableIterator = fieldTable.getIterator();
-        while (fieldTableIterator.hasNext()) {
-            setFieldValueToMQRFH2Header(fieldTableIterator, header);
-        }
-        return header;
-    }
-
-    private static void setFieldValueToMQRFH2Header(BIterator fieldTableIterator, MQRFH2 header) {
-        BMap<BString, Object> bField = (BMap<BString, Object>) ((BArray) fieldTableIterator.next()).get(1);
-        String folder = bField.getStringValue(FOLDER_FIELD).getValue();
-        String field = bField.getStringValue(FIELD_FIELD).getValue();
-        Object value = bField.get(VALUE_FIELD);
-        try {
-            if (value instanceof Long longValue) {
-                header.setLongFieldValue(folder, field, longValue.intValue());
-            } else if (value instanceof Integer intValue) {
-                header.setIntFieldValue(folder, field, intValue);
-            } else if (value instanceof Boolean booleanValue) {
-                header.setFieldValue(folder, field, booleanValue);
-            } else if (value instanceof Byte byteValue) {
-                header.setByteFieldValue(folder, field, byteValue);
-            } else if (value instanceof byte[] bytesValue) {
-                header.setFieldValue(folder, field, bytesValue);
-            } else if (value instanceof Float floatValue) {
-                header.setFloatFieldValue(folder, field, floatValue);
-            } else if (value instanceof Double doubleValue) {
-                header.setFloatFieldValue(folder, field, doubleValue.floatValue());
-            } else if (value instanceof BString stringValue) {
-                header.setFieldValue(folder, field, stringValue.getValue());
-            } else {
-                header.setFieldValue(folder, field, value);
-            }
-        } catch (IOException e) {
-            throw createError(IBMMQ_ERROR, String
-                    .format("Error occurred while setting field values to MQRFH2 header: %s", e.getMessage()), e);
         }
     }
 
@@ -331,7 +235,7 @@ public class CommonUtils {
     private static Object getBHeaders(Runtime runtime, MQMessage mqMessage) {
         ArrayList<BMap<BString, Object>> bHeaders = new ArrayList<>();
         try {
-            readHeadersFromMQMessage(runtime, bHeaders, HeaderType.MQRFH2, mqMessage);
+            MQRFH2Header.decodeHeader(runtime, mqMessage, bHeaders);
         } catch (IOException e) {
             throw createError(IBMMQ_ERROR,
                     String.format("Error occurred while reading headers: %s", e.getMessage()), e);
@@ -344,224 +248,6 @@ public class CommonUtils {
             headerArray.append(header);
         }
         return headerArray;
-    }
-
-    private static void readHeadersFromMQMessage(Runtime runtime, ArrayList<BMap<BString, Object>> headers,
-                                                 HeaderType type, MQMessage msg) throws IOException {
-        int dataOffset = msg.getDataOffset();
-        switch (type) {
-            case MQRFH2: {
-                MQRFH2 mqrfh2 = new MQRFH2();
-                try {
-                    mqrfh2.read(msg);
-                    headers.add(getBHeaderFromMQRFH2(runtime, mqrfh2));
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQDLH: {
-                MQDLH dlh = new MQDLH();
-                try {
-                    // Only MQRFH2 headers is supported at the moment. Other headers are read here
-                    // to move the cursor to the payload value.
-                    dlh.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQRFH: {
-                MQRFH mqrfh = new MQRFH();
-                try {
-                    mqrfh.read(msg);
-                    headers.add(getBHeaderFromMQRFH(mqrfh));
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQCIH: {
-                MQCIH mqcih = new MQCIH();
-                try {
-                    mqcih.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQIIH: {
-                MQIIH mqiih = new MQIIH();
-                try {
-                    mqiih.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQTM: {
-                MQTM mqtm = new MQTM();
-                try {
-                    mqtm.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQRMH: {
-                MQRMH mqrmh = new MQRMH();
-                try {
-                    mqrmh.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQSAPH: {
-                MQSAPH mqsaph = new MQSAPH();
-                try {
-                    mqsaph.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQWIH: {
-                MQWIH mqwih = new MQWIH();
-                try {
-                    mqwih.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQXQH: {
-                MQXQH mqxqh = new MQXQH();
-                try {
-                    mqxqh.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQDH: {
-                MQDH mqdh = new MQDH();
-                try {
-                    mqdh.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                }
-            }
-            /*-fallthrough*/
-            case MQEPH: {
-                MQEPH mqeph = new MQEPH();
-                try {
-                    mqeph.read(msg);
-                    break;
-                } catch (MQDataException e) {
-                    msg.seek(dataOffset);
-                    return;
-                }
-            }
-        }
-        readHeadersFromMQMessage(runtime, headers, HeaderType.MQRFH2, msg);
-    }
-
-    private static BMap<BString, Object> getBHeaderFromMQRFH2(Runtime runtime, MQRFH2 mqrfh2) throws IOException {
-        BMap<BString, Object> header = ValueCreator.createRecordValue(getModule(), MQRFH2_RECORD_NAME);
-        header.put(FLAGS_FIELD, mqrfh2.getFlags());
-        BArray folderStringArray = ValueCreator.createArrayValue(TypeCreator
-                .createArrayType(PredefinedTypes.TYPE_STRING));
-        String[] folderStrings = mqrfh2.getFolderStrings();
-        for (String folderString : folderStrings) {
-            folderStringArray.append(StringUtils.fromString(folderString));
-        }
-        header.put(FOLDER_STRINGS_FIELD, folderStringArray);
-        header.put(NAME_VALUE_CCSID_FIELD, mqrfh2.getNameValueCCSID());
-        header.put(NAME_VALUE_DATA_FIELD, ValueCreator.createArrayValue(mqrfh2.getNameValueData()));
-        header.put(STRUC_ID_FIELD, StringUtils.fromString(mqrfh2.getStrucId()));
-        header.put(STRUC_LENGTH_FIELD, mqrfh2.getStrucLength());
-        header.put(VERSION_FIELD, mqrfh2.getVersion());
-        BTable fieldValuesTable = getBHeaderFieldValuesFromMQMessage(runtime, mqrfh2);
-        header.put(FIELD_VALUES_FIELD, fieldValuesTable);
-        return header;
-    }
-
-    private static BMap<BString, Object> getBHeaderFromMQRFH(MQRFH mqrfh) {
-        BMap<BString, Object> header = ValueCreator.createRecordValue(getModule(), MQRFH_RECORD_NAME);
-        header.put(FLAGS_FIELD, mqrfh.getFlags());
-        header.put(STRUC_ID_FIELD, StringUtils.fromString(mqrfh.getStrucId()));
-        header.put(STRUC_LENGTH_FIELD, mqrfh.getStrucLength());
-        header.put(VERSION_FIELD, mqrfh.getVersion());
-        header.put(NAME_VALUE_PAIRS_FIELD, getBNameValuePairsFromMQRFH(mqrfh));
-        return header;
-    }
-
-    private static BMap<BString, Object> getBNameValuePairsFromMQRFH(MQRFH mqrfh) {
-        BMap<BString, Object> nameValuePairs = ValueCreator.createMapValue();
-        try {
-            for (MQRFH.NameValuePair nameValuePair : (List<MQRFH.NameValuePair>) mqrfh.getNameValuePairs()) {
-                nameValuePairs.put(StringUtils.fromString(nameValuePair.getName()),
-                        StringUtils.fromString(nameValuePair.getValue()));
-            }
-        } catch (IOException e) {
-            throw createError(IBMMQ_ERROR, String
-                    .format("Error occurred while adding name value pairs to MQRFH header: %s", e.getMessage()), e);
-        }
-        return nameValuePairs;
-    }
-
-    private static BTable getBHeaderFieldValuesFromMQMessage(Runtime runtime, MQRFH2 mqrfh2) throws IOException {
-        BArray fieldArray = ValueCreator.createArrayValue(TypeCreator.createArrayType(TypeCreator
-                .createRecordType(MQRFH2FIELD_RECORD_NAME, getModule(), 0, false, 0)));
-        MQRFH2.Element[] folders = mqrfh2.getFolders();
-        int i = 0;
-        for (MQRFH2.Element folder : folders) {
-            MQRFH2.Element[] children = folder.getChildren();
-            for (MQRFH2.Element child : children) {
-                BMap<BString, Object> field = ValueCreator.createRecordValue(getModule(), MQRFH2FIELD_RECORD_NAME);
-                field.put(FOLDER_FIELD, StringUtils.fromString(folder.getName()));
-                field.put(FIELD_FIELD, StringUtils.fromString(child.getName()));
-                field.put(VALUE_FIELD, getBValueForMQObjectValue(child.getValue()));
-                fieldArray.add(i, field);
-                i = i + 1;
-            }
-        }
-        BObject nativeUtilsObject = ValueCreator.createObjectValue(getModule(), NATIVE_UTILS_OBJECT_NAME);
-        CountDownLatch latch = new CountDownLatch(1);
-        HeaderFieldValuesCallback headerFieldValuesCallback = new HeaderFieldValuesCallback(latch);
-        runtime.invokeMethodAsyncConcurrently(nativeUtilsObject, ADD_FIELDS_TO_TABLE_FUNCTION_NAME, null,
-                null, headerFieldValuesCallback, null, PredefinedTypes.TYPE_ANY, fieldArray, true);
-        try {
-            latch.await();
-        } catch (InterruptedException e) {
-            throw createError(IBMMQ_ERROR,
-                    String.format("Error occurred while adding MQRFH2 fields: %s", e.getMessage()), e);
-        }
-        return headerFieldValuesCallback.getHeaderValueTable();
-    }
-
-    private static Object getBValueForMQObjectValue(Object value) {
-        if (value instanceof Integer intValue) {
-            return intValue.longValue();
-        } else if (value instanceof String stringValue) {
-            return StringUtils.fromString(stringValue);
-        } else if (value instanceof byte[] bytesValue) {
-            return ValueCreator.createArrayValue(bytesValue);
-        } else {
-            return value;
-        }
     }
 
     public static BError createError(String errorType, String message, Throwable throwable) {
