+++ spotbugs-exclude.xml
@@ -29,7 +29,7 @@
         <Bug pattern="EI_EXPOSE_REP"/>
     </Match>
     <Match>
-        <Class name="io.ballerina.lib.ibm.ibmmq.CommonUtils"/>
-        <Bug pattern="SF_SWITCH_FALLTHROUGH"/>
+        <Class name="io.ballerina.lib.ibm.ibmmq.headers.MQRFH2Header"/>
+        <Bug pattern="IL_INFINITE_RECURSIVE_LOOP"/>
     </Match>
 </FindBugsFilter>
